CU-24wgdcy factoryにフォルダを作成して、ディレクトリ構成を変更した
