+++ backend/database/factories/Blogs/BlogResponseGoodFactory.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Database\Factories;
+namespace Database\Factories\Blogs;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
