+++ backend/database/factories/Countries/FavoriteCountryFactory.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Database\Factories;
+namespace Database\Factories\Countries;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
