+++ backend/database/factories/Hitchhikes/Doings/DriverReviewFactory.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Database\Factories;
+namespace Database\Factories\Hitchhikes\Doings;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-class HitchhikeFactory extends Factory
+class DriverReviewFactory extends Factory
 {
     /**
      * Define the model's default state.
