+++ backend/database/factories/Hitchhikes/Doings/HitchhikeMessageFactory.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Database\Factories;
+namespace Database\Factories\Hitchhikes\Doings;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-class HikerReviewFactory extends Factory
+class HitchhikeMessageFactory extends Factory
 {
     /**
      * Define the model's default state.
