+++ backend/database/factories/Hitchhikes/Doings/PickUpFactory.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Database\Factories;
+namespace Database\Factories\Hitchhikes\Doings;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-class BlogFactory extends Factory
+class PickUpFactory extends Factory
 {
     /**
      * Define the model's default state.
