+++ backend/database/factories/Hitchhikes/Spots/HitchhikeFactory.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Database\Factories;
+namespace Database\Factories\Hitchhikes\Spots;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-class CountryFactory extends Factory
+class HitchhikeFactory extends Factory
 {
     /**
      * Define the model's default state.
