+++ backend/database/factories/Hitchhikes/Spots/HitchhikeGoodFactory.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Database\Factories;
+namespace Database\Factories\Hitchhikes\Spots;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-class MemoGoodFactory extends Factory
+class HitchhikeGoodFactory extends Factory
 {
     /**
      * Define the model's default state.
