+++ backend/database/factories/Hitchhikes/Spots/SuccessFactory.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Database\Factories;
+namespace Database\Factories\Hitchhikes\Spots;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-class RequestFactory extends Factory
+class SuccessFactory extends Factory
 {
     /**
      * Define the model's default state.
