+++ backend/database/factories/Memos/MemoResponseFactory.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Database\Factories;
+namespace Database\Factories\Memos;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-class HitchhikeGoodFactory extends Factory
+class MemoResponseFactory extends Factory
 {
     /**
      * Define the model's default state.
