+++ backend/database/factories/Memos/MemoResponseGoodFactory.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Database\Factories;
+namespace Database\Factories\Memos;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 
-class DriverReviewFactory extends Factory
+class MemoResponseGoodFactory extends Factory
 {
     /**
      * Define the model's default state.
