+++ backend/database/factories/Questions/QuestionFactory.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Database\Factories;
+namespace Database\Factories\Questions;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
